Resubmisson July 3

+++ 6_25_2024_parenthesization_pytorch/data.py
@@ -44,7 +44,7 @@ def parenthesization_to_tensor(parenthesization):
         for item in sublist:
             Flattened_ParenList.append(item)
         
-    return torch.tensor(Flattened_ParenList, dtype=torch.int8)
+    return torch.tensor(Flattened_ParenList, dtype= torch.float32)
 
 
 class ParenthesizationDataset(Dataset):
